+++ primitives/consensus/sassafras/src/lib.rs
@@ -165,10 +165,10 @@ pub fn make_slot_transcript_data(
 }
 
 /// Make ticket VRF transcript.
-pub fn make_ticket_transcript(randomness: &Randomness, attempt: u64, epoch: u64) -> Transcript {
+pub fn make_ticket_transcript(randomness: &Randomness, attempt: u32, epoch: u64) -> Transcript {
 	let mut transcript = Transcript::new(&SASSAFRAS_ENGINE_ID);
 	transcript.append_message(TYPE_LABEL.as_bytes(), TICKET_VRF_TYPE_VALUE.as_bytes());
-	transcript.append_u64(ATTEMPT_LABEL.as_bytes(), attempt);
+	transcript.append_u64(ATTEMPT_LABEL.as_bytes(), attempt as u64);
 	transcript.append_u64(EPOCH_LABEL.as_bytes(), epoch);
 	transcript.append_message(RANDOMNESS_LABEL.as_bytes(), randomness);
 	transcript
@@ -178,14 +178,14 @@ pub fn make_ticket_transcript(randomness: &Randomness, attempt: u64, epoch: u64)
 #[cfg(feature = "std")]
 pub fn make_ticket_transcript_data(
 	randomness: &Randomness,
-	attempt: u64,
+	attempt: u32,
 	epoch: u64,
 ) -> VRFTranscriptData {
 	VRFTranscriptData {
 		label: &SASSAFRAS_ENGINE_ID,
 		items: vec![
 			(TYPE_LABEL, VRFTranscriptValue::Bytes(TICKET_VRF_TYPE_VALUE.as_bytes().to_vec())),
-			(ATTEMPT_LABEL, VRFTranscriptValue::U64(attempt)),
+			(ATTEMPT_LABEL, VRFTranscriptValue::U64(attempt as u64)),
 			(EPOCH_LABEL, VRFTranscriptValue::U64(epoch)),
 			(RANDOMNESS_LABEL, VRFTranscriptValue::Bytes(randomness.to_vec())),
 		],
